Update models for changed EYFS task.
