+++ arc_application/models/__init__.py
@@ -10,6 +10,7 @@ from .arc import *
 from .arc_comments import *
 from .child_in_home import *
 from .childcare_type import *
+from .childcare_training import *
 from .criminal_record_check import *
 from .childcare_training import *
 from .firstaid_training import *
