+++ arc_application/models/application.py
@@ -44,8 +44,8 @@ class Application(models.Model):
     childcare_type_arc_flagged = models.BooleanField(default=False)
     first_aid_training_status = models.CharField(choices=TASK_STATUS, max_length=50)
     first_aid_training_arc_flagged = models.BooleanField(default=False)
-    eyfs_training_status = models.CharField(choices=TASK_STATUS, max_length=50)
-    eyfs_training_arc_flagged = models.BooleanField(default=False)
+    childcare_training_status = models.CharField(choices=TASK_STATUS, max_length=50)
+    childcare_training_arc_flagged = models.BooleanField(default=False)
     criminal_record_check_status = models.CharField(choices=TASK_STATUS, max_length=50)
     criminal_record_check_arc_flagged = models.BooleanField(default=False)
     health_status = models.CharField(choices=TASK_STATUS, max_length=50)
